Fixes rssi publishing

+++ bt-sensor/rssi.py
@@ -75,7 +75,7 @@ def talker(btad_list):
         for i,pub in enumerate(pub_list):
             rssi = btobj_list[i].get_rssi()
             rospy.loginfo(rssi)
-            pub.publish(rssi)
+            pub.publish(Int8(rssi))
             rate.sleep()
 
 if __name__ == '__main__':
